project/api_gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
slowapi==0.1.9
pydantic==2.5.0
python-dotenv==1.0.0
pino==0.2.0
fastapi-limiter==0.1.6
python-jose[cryptography]==3.3.0
redis==5.0.1